requirements.txt
fastapi
uvicorn
langchain-google-genai
langchain-chroma
chromadb
python-dotenv
rank_bm25
pydantic
sqlalchemy
tenacity
alembic
apscheduler
pytest
pytest-asyncio
httpx
arxiv
langchain-ollama
feedparser
markdown
slowapi
